+++ src/dropdown-multi-select/index.js
@@ -1,0 +1,1 @@
+export DropdownMultiSelect from './dropdown-multi-select.component.jsx';